each value knows its unit and scaler (only before calculation)
